+++ src/main/java/net/telepathicgrunt/bumblezone/entities/mobs/HoneySlimeEntity.java
@@ -39,10 +39,10 @@ import net.minecraft.world.LocalDifficulty;
 import net.minecraft.world.ServerWorldAccess;
 import net.minecraft.world.World;
 import net.minecraft.world.WorldAccess;
-import net.telepathicgrunt.bumblezone.modinit.BzBlocks;
-import net.telepathicgrunt.bumblezone.modinit.BzEntities;
 import net.telepathicgrunt.bumblezone.entities.controllers.HoneySlimeMoveHelperController;
 import net.telepathicgrunt.bumblezone.entities.goals.*;
+import net.telepathicgrunt.bumblezone.modinit.BzBlocks;
+import net.telepathicgrunt.bumblezone.modinit.BzEntities;
 
 import java.util.HashSet;
 import java.util.Objects;